fix: 스트레스 계산 알고리즘 수정

+++ backend/fast-api/models/request_models.py
@@ -6,5 +6,5 @@ class SingleStressRequest(BaseModel):
 
 
 class StressOverviewRequest(BaseModel):
-    hr_data: list[int]
+    hr_data: list[float]
     walking_time: int
